+++ src/components/UserDetail/UserDetail.js
@@ -19,6 +19,8 @@ import {
 } from '@folio/stripes/components';
 
 import { getFullName } from '../../util';
+import userPlaceholder from '../../../icons/user-placeholder.png';
+
 
 import css from './UserDetail.css';
 
@@ -200,10 +202,7 @@ class UserDetail extends React.Component {
             </Col>
             {hasProfilePicture &&
               <Col xs={2}>
-                <img
-                  src="http://placehold.it/60x60"
-                  alt={<FormattedMessage id="ui-checkout.presentation" />}
-                />
+                <img className={`floatEnd ${css.userPlaceholder}`} src={userPlaceholder} alt="presentation" />
               </Col> }
           </Row>
         </div>