+++ test/bigtest/interactors/block-modal.js
@@ -9,10 +9,6 @@ import {
 
   modalPresent = isPresent('[data-test-block-modal]');
   modalMessage = collection('[data-test-block-message]');
-
-  whenBlockMessageIsLoaded() {
-    return this.when(() => this.modalMessage.isPresent);
-  }
 }
 
 export default BlockModalInteractor;