+++ test/bigtest/interactors/block-modal.js
@@ -1,12 +1,14 @@
 import {
   interactor,
   isPresent,
+  collection,
 } from '@bigtest/interactor';
 
 @interactor class BlockModalInteractor {
   static defaultScope = 'body';
 
   modalPresent = isPresent('[data-test-block-modal]');
+  modalMessage = collection('[data-test-block-message]');
 }
 
 export default BlockModalInteractor;