+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade to `stripes` `4.0`, `react-intl` `4.5`. Refs STRIPES-672.
 * Upgrade to `react-intl-safe-html` `2.0`. Refs STRIPES-672.
+* Use local file as a profile placeholder. Refs UICHKOUT-624
 
 ## [3.1.0] (IN PROGRESS)