+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Upgrade to `stripes` `4.0`, `react-intl` `4.5`. Refs STRIPES-672.
 * Upgrade to `react-intl-safe-html` `2.0`. Refs STRIPES-672.
 * Use local file as a profile placeholder. Refs UICHKOUT-624.
+* Add confirmation modal for Lost and paid items. Refs UICHKOUT-549.
 * Include automated patron blocks as reason to block patron from borrowing. Refs UICHKOUT-627.
 
 ## [3.1.0] (IN PROGRESS)