+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Upgrade to `react-intl-safe-html` `2.0`. Refs STRIPES-672.
 * Use local file as a profile placeholder. Refs UICHKOUT-624.
 * Add confirmation modal for Lost and paid items. Refs UICHKOUT-549.
+* Include automated patron blocks as reason to block patron from borrowing. Refs UICHKOUT-627.
 
 ## [3.1.0] (IN PROGRESS)